src/main/java/ru/job4j/io/Config.java
package ru.job4j.io;

import org.w3c.dom.ls.LSOutput;

import java.io.BufferedReader;
import java.io.FileReader;
import java.io.IOException;
import java.util.Arrays;
import java.util.HashMap;
import java.util.Map;
import java.util.StringJoiner;
import java.util.stream.Collectors;

public class Config {

    private final String path;
    private final Map<String, String> values = new HashMap<String, String>();

    public Config(final String path) {
        this.path = path;
    }

    public void load() {
        for (String s : Arrays.stream(toString()
                .split(System.lineSeparator()))
                .filter(a -> !a.startsWith("#"))
                .filter(a -> !a.isEmpty())
                .filter(a -> a.contains("="))
                .collect(Collectors.toList())) {
            String key = s.substring(0, s.indexOf("="));
            String value = s.substring(s.indexOf("=") + 1);
            if (key.isEmpty() || value.isEmpty()) {
                throw new IllegalArgumentException("error");
            }
            values.put(key, value);
        }
    }

    public String value(String key) {
        return values.get(key);
    }

    @Override
    public String toString() {
        StringJoiner out = new StringJoiner(System.lineSeparator());
        try (BufferedReader read = new BufferedReader(new FileReader(this.path))) {
            read.lines().forEach(out::add);
        } catch (IOException e) {
            e.printStackTrace();
        }
        return out.toString();
    }

}